apply API to register custom components
* apply API to register custom components
  * https://github.com/plataformatec/simple_form/pull/1556
