+++ config/initializers/simple_form_bootstrap.rb
@@ -381,27 +381,5 @@ SimpleForm.setup do |config|
 end
 
 
-# input group
-#
-# custom component requires input group wrapper
-module SimpleForm
-  module Components
-    module InputGroups
-      module Prepends
-        def prepend(wrapper_options = nil)
-          span_tag = content_tag(:span, options[:prepend], class: "input-group-text")
-          template.content_tag(:div, span_tag, class: "input-group-prepend")
-        end
-      end
-
-      module Appends
-        def append(wrapper_options = nil)
-          span_tag = content_tag(:span, options[:append], class: "input-group-text")
-          template.content_tag(:div, span_tag, class: "input-group-append")
-        end
-      end
-    end
-  end
-end
-SimpleForm::Inputs::Base.send(:include, SimpleForm::Components::InputGroups::Prepends)
-SimpleForm::Inputs::Base.send(:include, SimpleForm::Components::InputGroups::Appends)
+# load custom components
+Dir[Rails.root.join('app/components/**/*.rb')].each { |f| require f }
